html/semantics/embedded-content-0/media-elements/historical.html

<!doctype html>
<title>Historical media element features should not be supported</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
// <video audio=muted>: added in r5636, replaced with <video muted> in r5991.
test(function() {
  assert_false('audio' in document.createElement('video'));
}, '<video audio=muted> should not be supported')
</script>
